pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = "Sample Python project with MySQL and Flask using Poetry"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^8.0"
flask = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
